requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.3.2=pypi_0
aiohttp=3.10.0=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py310hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.4=h56e8100_0
certifi=2024.7.4=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
colour=0.1.5=pypi_0
contourpy=1.2.1=pypi_0
cuda-cccl=12.5.39=0
cuda-cccl_win-64=12.5.39=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-opencl-dev=12.5.39=0
cuda-profiler-api=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
filelock=3.13.1=py310haa95532_0
fonttools=4.53.1=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
gmpy2=2.1.2=py310h7f96b67_0
grpcio=1.65.2=pypi_0
idna=3.7=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.4=py310haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_2
kiwisolver=1.4.5=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.6.82=0
libcurand-dev=10.3.6.82=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h2bbff1b_0
lightning-utilities=0.11.6=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markdown=3.6=pypi_0
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib=3.9.1=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
multidict=6.0.5=pypi_0
networkx=3.3=py310haa95532_0
numpy=1.26.4=py310h055cbcc_0
numpy-base=1.26.4=py310h65a83cf_0
opencv-python=4.10.0.84=pypi_0
openjpeg=2.4.0=h4afccc4_2
openssl=3.3.1=h2466b09_2
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.4.0=py310h827c3e9_0
pip=24.0=py310haa95532_0
protobuf=4.25.4=pypi_0
psutil=6.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyexiv2=2.14.0=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.12=he1021f5_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.1.2=py3.10_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-lightning=2.3.3=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h2bbff1b_0
requests=2.32.3=py310haa95532_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=69.5.1=py310haa95532_0
shapely=2.0.5=pypi_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.12=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
torchaudio=2.1.2=pypi_0
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.16.2=pypi_0
tqdm=4.66.4=pypi_0
typing_extensions=4.11.0=py310haa95532_0
tzdata=2024.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
ultralytics=8.2.70=pypi_0
ultralytics-thop=2.0.0=pypi_0
urllib3=2.2.2=py310haa95532_0
vc=14.2=h2eaa2aa_4
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
